feat: Move HexEncodedAmount out to renamed hexEncodedTypes.ts and reuse Update imports
