+++ packages/iota/src/models/IBlock.ts
@@ -1,6 +1,6 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-import type { HexEncodedString } from "./hexEncodedString";
+import type { HexEncodedString } from "./hexEncodedTypes";
 import type { IMilestonePayload } from "./payloads/IMilestonePayload";
 import type { ITaggedDataPayload } from "./payloads/ITaggedDataPayload";
 import type { ITransactionPayload } from "./payloads/ITransactionPayload";
