+++ packages/iota/src/models/IBlockMetadata.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 import type { ConflictReason } from "./conflictReason";
-import type { HexEncodedString } from "./hexEncodedString";
+import type { HexEncodedString } from "./hexEncodedTypes";
 import type { LedgerInclusionState } from "./ledgerInclusionState";
 
 /**
