+++ packages/iota/src/models/IMigratedFunds.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 import type { AddressTypes } from "../models/addresses/addressTypes";
-import type { HexEncodedString } from "./hexEncodedString";
+import type { HexEncodedString } from "./hexEncodedTypes";
 
 /**
  * The migrated funds for receipts.
