+++ packages/iota/src/models/api/IOutputMetadataResponse.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import type { HexEncodedString } from "../hexEncodedString";
+import type { HexEncodedString } from "../hexEncodedTypes";
 
 /**
  * Details of an output.
