+++ packages/iota/src/models/hexEncodedTypes.ts
@@ -2,6 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * Hex encoded U256.
+ * Hex encoded bytes.
  */
 export type HexEncodedString = string;
+
+/**
+ * Hex encoded bytes.
+ */
+export type HexEncodedAmount = string;
+
