+++ packages/iota/src/models/outputs/IAliasOutput.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 import type { FeatureTypes } from "../features/featureTypes";
-import type { HexEncodedString } from "../hexEncodedString";
+import type { HexEncodedString } from "../hexEncodedTypes";
 import type { ITypeBase } from "../ITypeBase";
 import type { ICommonOutput } from "./ICommonOutput";
