+++ packages/iota/src/models/tokenSchemes/ISimpleTokenScheme.ts
@@ -1,6 +1,6 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-import type { HexEncodedString } from "../hexEncodedString";
+import type { HexEncodedAmount } from "../hexEncodedTypes";
 import type { ITypeBase } from "../ITypeBase";
 
 /**
@@ -15,15 +15,15 @@ export interface ISimpleTokenScheme extends ITypeBase<0> {
     /**
      * Amount of tokens minted by this foundry.
      */
-    mintedTokens: HexEncodedString;
+    mintedTokens: HexEncodedAmount;
 
     /**
      * Amount of tokens melted by this foundry.
      */
-    meltedTokens: HexEncodedString;
+    meltedTokens: HexEncodedAmount;
 
     /**
      * Maximum supply of tokens controlled by this foundry.
      */
-    maximumSupply: HexEncodedString;
+    maximumSupply: HexEncodedAmount;
 }
